simplify internal image state
