+++ mapillary_tools/process_sequence_properties.py
@@ -3,6 +3,7 @@ import datetime
 import os
 import uuid
 
+from . import image_log
 from . import processing, types
 from .geo import compute_bearing, gps_distance, diff_bearing, gps_speed
 
@@ -10,12 +11,120 @@ MAX_SEQUENCE_LENGTH = 500
 MAX_CAPTURE_SPEED = 45  # in m/s
 
 
+def load_geotag_points(
+    process_file_list: T.List[str],
+) -> T.Tuple[
+    T.List[str], T.List[datetime.datetime], T.List[float], T.List[float], T.List[float]
+]:
+    file_list = []
+    capture_times = []
+    lats = []
+    lons = []
+    directions = []
+
+    for image in process_file_list:
+        ret = image_log.read_process_data_from_memory(image, "geotag_process")
+        if ret is None:
+            continue
+        status, geotag_data = ret
+        if status != "success":
+            continue
+
+        # assume all data needed available from this point on
+        file_list.append(image)
+        capture_times.append(
+            datetime.datetime.strptime(
+                geotag_data["MAPCaptureTime"], "%Y_%m_%d_%H_%M_%S_%f"
+            )
+        )
+        lats.append(geotag_data["MAPLatitude"])
+        lons.append(geotag_data["MAPLongitude"])
+        directions.append(
+            geotag_data["MAPCompassHeading"]["TrueHeading"]
+        ) if "MAPCompassHeading" in geotag_data else directions.append(0.0)
+
+        image_log.unmark_duplicated(image)
+
+    return file_list, capture_times, lats, lons, directions
+
+
+def split_sequences(
+    capture_times: T.List[datetime.datetime],
+    lats: T.List[float],
+    lons: T.List[float],
+    file_list: T.List[str],
+    directions: T.List[float],
+    cutoff_time: float,
+    cutoff_distance: float,
+) -> T.List[T.Dict]:
+    sequences: T.List[T.Dict] = []
+    # sort based on time
+    sort_by_time = list(zip(capture_times, file_list, lats, lons, directions))
+    sort_by_time.sort()
+    capture_times, file_list, lats, lons, directions = [
+        list(x) for x in zip(*sort_by_time)
+    ]
+    latlons = list(zip(lats, lons))
+
+    # initialize first sequence
+    sequence_index = 0
+    sequences.append(
+        {
+            "file_list": [file_list[0]],
+            "directions": [directions[0]],
+            "latlons": [latlons[0]],
+            "capture_times": [capture_times[0]],
+        }
+    )
+
+    if len(file_list) >= 1:
+        # diff in capture time
+        capture_deltas = [t2 - t1 for t1, t2 in zip(capture_times, capture_times[1:])]
+
+        # distance between consecutive images
+        distances = [gps_distance(ll1, ll2) for ll1, ll2 in zip(latlons, latlons[1:])]
+
+        # if cutoff time is given use that, else assume cutoff is
+        # 1.5x median time delta
+        if cutoff_time is None:
+            median = sorted(capture_deltas)[len(capture_deltas) // 2]
+            if type(median) is not int:
+                median = median.total_seconds()
+            cutoff_time = 1.5 * median
+        else:
+            cutoff_time = float(cutoff_time)
+        cut = 0
+        for i, filepath in enumerate(file_list[1:]):
+            cut_time = capture_deltas[i].total_seconds() > cutoff_time
+            cut_distance = distances[i] > cutoff_distance
+            if cut_time or cut_distance:
+                cut += 1
+                # delta too big, start new sequence
+                sequence_index += 1
+                sequences.append(
+                    {
+                        "file_list": [filepath],
+                        "directions": [directions[1:][i]],
+                        "latlons": [latlons[1:][i]],
+                        "capture_times": [capture_times[1:][i]],
+                    }
+                )
+            else:
+                # delta not too big, continue with current
+                # group
+                sequences[sequence_index]["file_list"].append(filepath)
+                sequences[sequence_index]["directions"].append(directions[1:][i])
+                sequences[sequence_index]["latlons"].append(latlons[1:][i])
+                sequences[sequence_index]["capture_times"].append(capture_times[1:][i])
+
+    return sequences
+
+
 def finalize_sequence_processing(
     sequence,
     final_file_list,
     final_directions,
     final_capture_times,
-    verbose=False,
 ):
     for image, direction, capture_time in zip(
         final_file_list, final_directions, final_capture_times
@@ -30,8 +139,8 @@ def finalize_sequence_processing(
                 capture_time, "%Y_%m_%d_%H_%M_%S_%f"
             )[:-3],
         }
-        processing.create_and_log_process(
-            image, "sequence_process", "success", mapillary_description, verbose=verbose
+        image_log.create_and_log_process_in_memory(
+            image, "sequence_process", "success", mapillary_description
         )
 
 
@@ -44,7 +153,6 @@ def process_sequence_properties(
     duplicate_distance=0.1,
     duplicate_angle=5,
     offset_angle=0.0,
-    verbose=False,
     rerun=False,
     skip_subfolders=False,
 ):
@@ -53,7 +161,7 @@ def process_sequence_properties(
         raise RuntimeError(f"Error, import directory {import_path} does not exist")
 
     sequences = find_sequences(
-        cutoff_distance, cutoff_time, import_path, rerun, skip_subfolders, verbose
+        cutoff_distance, cutoff_time, import_path, rerun, skip_subfolders
     )
 
     # process for each sequence
@@ -123,7 +231,7 @@ def process_sequence_properties(
                     direction_diff = 360
 
                 if distance < duplicate_distance and direction_diff < duplicate_angle:
-                    processing.mark_as_duplicated(filename)
+                    image_log.mark_as_duplicated(filename)
                     # FIXME: understand why
                     # log_root = uploader.log_rootpath(filename)
                     # sequence_process_success_path = os.path.join(
@@ -151,12 +259,11 @@ def process_sequence_properties(
                 final_file_list[i : i + MAX_SEQUENCE_LENGTH],
                 final_directions[i : i + MAX_SEQUENCE_LENGTH],
                 final_capture_times[i : i + MAX_SEQUENCE_LENGTH],
-                verbose,
             )
 
 
 def find_sequences(
-    cutoff_distance, cutoff_time, import_path, rerun, skip_subfolders, verbose
+    cutoff_distance, cutoff_time, import_path, rerun, skip_subfolders
 ) -> T.List[T.Dict]:
     def _split(process_file_list: T.List[str]) -> T.List[T.Dict]:
         if not process_file_list:
@@ -168,10 +275,10 @@ def find_sequences(
             lats,
             lons,
             directions,
-        ) = processing.load_geotag_points(process_file_list, verbose)
+        ) = load_geotag_points(process_file_list)
 
         if capture_times and lats and lons:
-            return processing.split_sequences(
+            return split_sequences(
                 capture_times,
                 lats,
                 lons,
@@ -179,15 +286,14 @@ def find_sequences(
                 directions,
                 cutoff_time,
                 cutoff_distance,
-                verbose,
             )
         else:
             return []
 
     if skip_subfolders:
-        process_file_list = processing.get_process_file_list(
+        process_file_list = image_log.get_process_file_list(
             import_path,
-            "sequence_process",
+            "mapillary_image_description",
             rerun=rerun,
             skip_subfolders=True,
         )
@@ -198,9 +304,9 @@ def find_sequences(
         # sequence limited to the root of the files
         for root, dirs, files in os.walk(import_path, topdown=True):
             dirs[:] = [d for d in dirs if not d.startswith(".")]
-            process_file_list = processing.get_process_file_list(
+            process_file_list = image_log.get_process_file_list(
                 root,
-                "sequence_process",
+                "mapillary_image_description",
                 rerun=rerun,
                 skip_subfolders=True,
             )
